update: auto expend when not empty in staff accordion
